requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# YOLO + Torch CPU
ultralytics==8.3.1

# Bildverarbeitung
pillow==10.3.0
numpy==1.26.4
